Start betterpad with at least 80 characters in width
The chosen method for calculating width is the ugliest of all
the options available on .NET, but it is also the most accurate.

Other options that do not involve creating/using an actual on-
screen object end up getting the wrong device context and wind up
very, very wrong on hi-dpi devices.