Implemented smooth scroll replacement, but not activated
After actually disabling the smooth scrolling and replacing it
with a fixed 3-line scroll per mouse wheel indent, I'm not sure
I like the result, so it is disabled for now. The code remains
should there be a need for a validated solution to disabling
smooth scroll at some point in the future.