Refactor recovery to handle power loss, etc
The old behavior relied on randomly-generated paths which would
be supplied to the application by the MS Windows recovery API on
restart in case of system upgrade or a crash, and storing in the
unsafe-shutdown directory on QUERYENDSESSION to handle system
shutdown.

This approach ended up causing the only backup in the event of a
power loss to be the random directory, which would not have been
supplied to the application at next start. Now using the unsafe
shutdown directory always.