Fix using non-basic font weights
The `FontDialog` from `System.Windows.Forms` has broken support for
anything other than the most basic font weights, e.g. it tries to load
"Foo Sans Medium" as font family "Foo Sans Medium" with a font weight of
"Bold" rather than "Foo Sans" with font weight "Medium".

If the `FontFamily` as loaded by the `FontDialog` contains the weight,
assume that the correct weight in the `FontStyle` property is `Regular`,
regardless of what it is currently set to.